BDChat Beta2.2/BDChat Beta2.2/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
azure-ai-inference==1.0.0b9
azure-core>=1.29.5 